+++ native/src/main/java/io/ballerina/stdlib/websocket/WebSocketResourceCallback.java
@@ -19,8 +19,11 @@ package io.ballerina.stdlib.websocket;
 
 import io.ballerina.runtime.api.Runtime;
 import io.ballerina.runtime.api.concurrent.StrandMetadata;
+import io.ballerina.runtime.api.types.Type;
+import io.ballerina.runtime.api.utils.TypeUtils;
 import io.ballerina.runtime.api.values.BArray;
 import io.ballerina.runtime.api.values.BError;
+import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.runtime.api.values.BStream;
 import io.ballerina.runtime.api.values.BString;
@@ -39,11 +42,19 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.nio.ByteBuffer;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
+import java.util.concurrent.CountDownLatch;
 
 import static io.ballerina.runtime.api.utils.StringUtils.fromString;
+import static io.ballerina.stdlib.websocket.WebSocketConstants.CLOSE_FRAME_DEFAULT_TIMEOUT;
+import static io.ballerina.stdlib.websocket.WebSocketConstants.CLOSE_FRAME_TYPE;
+import static io.ballerina.stdlib.websocket.WebSocketConstants.PACKAGE_WEBSOCKET;
 import static io.ballerina.stdlib.websocket.WebSocketConstants.STREAMING_NEXT_FUNCTION;
 import static io.ballerina.stdlib.websocket.WebSocketResourceDispatcher.dispatchOnError;
+import static io.ballerina.stdlib.websocket.actions.websocketconnector.Close.initiateConnectionClosure;
+import static io.ballerina.stdlib.websocket.actions.websocketconnector.Close.waitForTimeout;
 import static io.ballerina.stdlib.websocket.actions.websocketconnector.WebSocketConnector.fromByteArray;
 import static io.ballerina.stdlib.websocket.actions.websocketconnector.WebSocketConnector.fromText;
 import static io.ballerina.stdlib.websocket.actions.websocketconnector.WebSocketConnector.getByteChunk;
@@ -68,6 +79,23 @@ public final class WebSocketResourceCallback implements Handler {
         this.connectionInfo = webSocketConnectionInfo;
         this.webSocketConnection = connectionInfo.getWebSocketConnection();
         this.resource = resource;
+    }
+
+    @SuppressWarnings(WebSocketConstants.UNCHECKED)
+    public static boolean isCloseFrameRecord(Object obj) {
+        if (obj instanceof BMap) {
+            BMap<BString, Object> bMap = (BMap<BString, Object>) obj;
+            if (bMap.containsKey(fromString(CLOSE_FRAME_TYPE)) &&
+                    bMap.get(fromString(CLOSE_FRAME_TYPE)) instanceof BObject) {
+                Type objectType = TypeUtils.getType(bMap.get(fromString(CLOSE_FRAME_TYPE)));
+                String objectName = objectType.getName();
+                String objectPackage = objectType.getPackage().getName();
+                return objectPackage.equals(PACKAGE_WEBSOCKET) &&
+                        (objectName.equals(WebSocketConstants.PREDEFINED_CLOSE_FRAME_TYPE) ||
+                                objectName.equals(WebSocketConstants.CUSTOM_CLOSE_FRAME_TYPE));
+            }
+        }
+        return false;
     }
 
     @Override
@@ -95,6 +123,8 @@ public final class WebSocketResourceCallback implements Handler {
                     returnStreamUnitCallBack.notifyFailure(bError);
                 }
             });
+        } else if (isCloseFrameRecord(result)) {
+            sendCloseFrame(result, connectionInfo);
         } else if (result == null) {
             webSocketConnection.readNextFrame();
         } else if (result instanceof BXml bXml) {
@@ -107,6 +137,41 @@ public final class WebSocketResourceCallback implements Handler {
         } else {
             log.error("invalid return type");
         }
+    }
+
+    public static void sendCloseFrame(Object result, WebSocketConnectionInfo connectionInfo) {
+        @SuppressWarnings(WebSocketConstants.UNCHECKED)
+        BMap<BString, Object> closeFrameRecord = (BMap<BString, Object>) result;
+        int statusCode = closeFrameRecord.getIntValue(WebSocketConstants.CLOSE_FRAME_STATUS_CODE).intValue();
+        String reason = closeFrameRecord.containsKey(WebSocketConstants.CLOSE_FRAME_REASON) ?
+                closeFrameRecord.getStringValue(WebSocketConstants.CLOSE_FRAME_REASON).getValue() : "";
+        try {
+            if (isValidStatusCode(statusCode)) {
+                CountDownLatch countDownLatch = new CountDownLatch(1);
+                List<BError> errors = new ArrayList<>(1);
+                ChannelFuture closeFuture = initiateConnectionClosure(errors, statusCode, reason,
+                        connectionInfo, countDownLatch);
+                connectionInfo.getWebSocketConnection().readNextFrame();
+                waitForTimeout(errors, CLOSE_FRAME_DEFAULT_TIMEOUT, countDownLatch, connectionInfo);
+                closeFuture.channel().close().addListener(future -> {
+                    WebSocketUtil.setListenerOpenField(connectionInfo);
+                });
+            } else {
+                log.error("Failed to send close frame. Invalid status code: {}", statusCode);
+                connectionInfo.getWebSocketConnection().getChannel().close().addListener(future -> {
+                    WebSocketUtil.setListenerOpenField(connectionInfo);
+                });
+            }
+        } catch (Exception e) {
+            log.error("Error occurred when sending close frame", e);
+            dispatchOnError(connectionInfo, e,
+                    connectionInfo.getWebSocketEndpoint().get(WebSocketConstants.INITIALIZED_BY_SERVICE).equals(true));
+        }
+    }
+
+    private static boolean isValidStatusCode(int statusCode) {
+        return !(statusCode <= 999 || statusCode >= 1004 && statusCode <= 1006 || statusCode >= 1012 &&
+                statusCode <= 2999 || statusCode > 4999);
     }
 
     private void sendPong(BArray result, PromiseCombiner promiseCombiner) {